requirements.txt
python-telegram-bot==20.7
requests==2.31.0
websocket-client==1.7.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
mplfinance==0.12.10b0
Pillow==10.1.0
SQLAlchemy==2.0.23
pytz==2023.3
python-dotenv==1.0.0
aiohttp==3.9.1
aiohttp
matplotlib
mplfinance
numpy
pandas
Pillow
python-dotenv
python-telegram-bot
pytz
requests
SQLAlchemy
websocket-client
